refactor: split tree algorithm file into manageable pieces
