+++ packages/web/src/algorithms/tree/enums.ts
@@ -1,0 +1,9 @@
+export enum NodeType {
+  New = 'New',
+  Reference = 'Reference',
+}
+
+export enum QCStatusType {
+  Pass = 'Pass',
+  Fail = 'Fail',
+}
