+++ packages/web/src/algorithms/tree/types.ts
@@ -1,0 +1,8 @@
+import { Nucleotide } from 'src/algorithms/types'
+import { AuspiceTreeNode } from 'auspice'
+
+export type MutationMap = Map<number, Nucleotide>
+
+export interface AuspiceTreeNodeExtended extends AuspiceTreeNode {
+  mutations?: MutationMap
+}
