+++ packages/web/src/state/algorithm/algorithm.sagas.ts
@@ -10,14 +10,15 @@ import type { AuspiceTreeNode } from 'auspice'
 import { changeColorBy } from 'auspice/src/actions/colors'
 
 import type { AnalysisParams, AnalysisResult, AnalysisResultWithoutClade } from 'src/algorithms/types'
-import type { FinalizeTreeParams, LocateInTreeParams } from 'src/algorithms/tree/locateInTree'
+import type { LocateInTreeParams } from 'src/algorithms/tree/treeFindNearestNodes'
+import type { FinalizeTreeParams } from 'src/algorithms/tree/treeAttachNodes'
 import type { QCResult, QCRulesConfig, RunQCParams } from 'src/algorithms/QC/runQC'
 import type { WorkerPools } from 'src/workers/types'
 import type { ParseThread } from 'src/workers/worker.parse'
 import type { AnalyzeThread } from 'src/workers/worker.analyze'
-import type { TreeBuildThread } from 'src/workers/worker.treeBuild'
+import type { TreeBuildThread } from 'src/workers/worker.treeFindNearest'
 import type { RunQcThread } from 'src/workers/worker.runQc'
-import type { TreeFinalizeThread } from 'src/workers/worker.treeFinalize'
+import type { TreeFinalizeThread } from 'src/workers/worker.treeAttachNodes'
 
 import { safeZip } from 'src/helpers/safeZip'
 import { notUndefined } from 'src/helpers/notUndefined'
